Add InitEnv for testing
